+++ components/bsb/timer/background/impl/src/androidMain/kotlin/com/flipperdevices/bsb/timer/background/service/TimerForegroundService.kt
@@ -10,7 +10,6 @@ import com.flipperdevices.bsb.timer.background.api.TimerApi
 import com.flipperdevices.bsb.timer.background.api.TimerStateListener
 import com.flipperdevices.bsb.timer.background.api.TimerTimestamp
 import com.flipperdevices.bsb.timer.background.di.ServiceDIComponent
-import com.flipperdevices.bsb.timer.background.model.ControlledTimerState
 import com.flipperdevices.core.di.ComponentHolder
 import com.flipperdevices.core.ktx.android.toFullString
 import com.flipperdevices.core.ktx.common.FlipperDispatchers
