+++ components/bsb/timer/background/impl/src/commonMain/kotlin/com/flipperdevices/bsb/timer/background/api/CommonTimerApi.kt
@@ -46,7 +46,6 @@ class CommonTimerApi(
         compositeListeners.removeListener(listener)
     }
 
-
     override fun setTimestampState(state: TimerTimestamp?) {
         scope.launch {
             if (state == null) {
