+++ components/bsb/timer/background/impl/src/commonMain/kotlin/com/flipperdevices/bsb/timer/background/api/delegates/TimerLoopJob.kt
@@ -1,9 +1,9 @@
 package com.flipperdevices.bsb.timer.background.api.delegates
 
 import com.flipperdevices.bsb.timer.background.api.TimerTimestamp
+import com.flipperdevices.bsb.timer.background.api.util.toState
 import com.flipperdevices.bsb.timer.background.model.ControlledTimerState
 import com.flipperdevices.bsb.timer.background.model.isOnPause
-import com.flipperdevices.bsb.timer.background.util.toState
 import com.flipperdevices.core.ktx.common.withLock
 import com.flipperdevices.core.log.LogTagProvider
 import kotlinx.coroutines.CoroutineScope
